requirements.txt
# Drywall QA — Prompted Segmentation
# Python 3.9+  |  CUDA 11.8+ recommended

# ── Core ML ──────────────────────────────────────────────────
torch>=2.0.0
torchvision>=0.15.0
transformers>=4.35.0          # CLIPSeg model + processor
accelerate>=0.24.0            # HuggingFace training utilities

# ── Vision + Augmentation ────────────────────────────────────
opencv-python>=4.8.0          # image loading, morphological ops, CCA
opencv-contrib-python>=4.8.0  # ximgproc thinning for clDice skeleton
Pillow>=10.0.0                # PIL for image loading in inference
albumentations>=1.3.1         # fast joint image+mask augmentation
scipy>=1.11.0                 # skeletonisation fallback for clDice

# ── Data ─────────────────────────────────────────────────────
roboflow>=1.1.0               # Roboflow dataset downloader
numpy>=1.24.0
PyYAML>=6.0

# ── Evaluation ───────────────────────────────────────────────
# torchmetrics is optional — we implement metrics from scratch
# torchmetrics>=1.2.0

# ── Logging + Utilities ──────────────────────────────────────
tqdm>=4.66.0                  # progress bars

# ── Report generation ────────────────────────────────────────
matplotlib>=3.7.0             # visualisation
seaborn>=0.12.0               # metric tables and plots
fpdf2>=2.7.0                  # PDF report generation (scripts/generate_report.py)

# ── Optional: SAM refinement ─────────────────────────────────
# Uncomment if using SAM-based mask refinement
# segment-anything @ git+https://github.com/facebookresearch/segment-anything.git

# ── Development ──────────────────────────────────────────────
pytest>=7.4.0
black>=23.0.0